feat: add Endpoints.Cache cachex and rename proc caching to Endpoints.ResultsCache
