+++ lib/logflare/endpoints/resolver.ex
@@ -2,7 +2,7 @@ defmodule Logflare.Endpoints.Resolver do
   @moduledoc """
   Finds or spawns Endpoint.Cache processes across the cluster. Unique process for query plus query params.
   """
-  alias Logflare.Endpoints.Cache
+  alias Logflare.Endpoints.ResultsCache
 
   require Logger
 
@@ -10,7 +10,7 @@ defmodule Logflare.Endpoints.Resolver do
   Lists all caches for an endpoint
   """
   def list_caches(%Logflare.Endpoints.Query{id: id}) do
-    endpoints = Cache.endpoints_part(id)
+    endpoints = ResultsCache.endpoints_part(id)
 
     :syn.members(endpoints, id)
     |> Enum.map(fn {pid, _} -> pid end)
@@ -21,7 +21,7 @@ defmodule Logflare.Endpoints.Resolver do
   Returns the resolved pid.
   """
   def resolve(%Logflare.Endpoints.Query{id: id} = query, params) do
-    endpoints = Cache.endpoints_part(query.id, params)
+    endpoints = ResultsCache.endpoints_part(query.id, params)
 
     {:via, :syn, {endpoints, {query.id, params}}}
     |> GenServer.whereis()
@@ -30,12 +30,12 @@ defmodule Logflare.Endpoints.Resolver do
         pid
 
       nil ->
-        spec = {Cache, {query, params}}
+        spec = {ResultsCache, {query, params}}
         Logger.debug("Starting up Endpoint.Cache for Endpoint.Query id=#{id}", endpoint_id: id)
 
         via =
           {:via, PartitionSupervisor,
-           {Logflare.Endpoints.Cache.PartitionSupervisor, {id, params}}}
+           {Logflare.Endpoints.ResultsCache.PartitionSupervisor, {id, params}}}
 
         case DynamicSupervisor.start_child(via, spec) do
           {:ok, pid} -> pid
