+++ lib/logflare_web/controllers/plugs/fetch_resource.ex
@@ -67,10 +67,10 @@ defmodule LogflareWeb.Plugs.FetchResource do
     endpoint =
       case is_uuid?(token_or_name) do
         true ->
-          Endpoints.get_query_by_token(token_or_name)
+          Endpoints.Cache.get_by(token: token_or_name, user_id: user_id)
 
         false when user_id != nil ->
-          Endpoints.get_by(name: token_or_name, user_id: user_id)
+          Endpoints.Cache.get_by(name: token_or_name, user_id: user_id)
 
         _ ->
           nil
@@ -84,7 +84,7 @@ defmodule LogflareWeb.Plugs.FetchResource do
           conn,
         _opts
       ) do
-    endpoint = Endpoints.get_by(name: name, user_id: user_id)
+    endpoint = Endpoints.Cache.get_by(name: name, user_id: user_id)
     assign(conn, :endpoint, endpoint)
   end
